update report cluster resource api
